refactor : 멀티모듈환경에서 Fixture를 공용모듈에서 관리하도록 변경
