+++ batch/src/test/java/org/meotppo/webti/job/StatisticJobNewTest.java
@@ -9,7 +9,6 @@ import static org.meotppo.webti.job.StatisticJobConfig.STATISTIC_JOB_NEW;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.meotppo.webti.config.MongoTest;
 import org.meotppo.webti.domain.entity.jpa.profile.Profile;
 import org.meotppo.webti.domain.entity.jpa.result.Statistic;
 import org.meotppo.webti.domain.entity.mongo.result.TestResult;
@@ -17,6 +16,7 @@ import org.meotppo.webti.domain.entity.type.MbtiType;
 import org.meotppo.webti.domain.repository.jpa.developertype.ProfileRepository;
 import org.meotppo.webti.domain.repository.jpa.statistic.StatisticRepository;
 import org.meotppo.webti.domain.repository.mongo.testresult.TestResultRepository;
+import org.meotppo.webti.job.config.MongoTest;
 import org.meotppo.webti.util.DatabaseUtils;
 import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.Job;
