+++ domain/build.gradle
@@ -1,5 +1,8 @@
 plugins {
     id 'java'
+    id 'java-library'
+    id 'java-test-fixtures'
+    id 'maven-publish'
     id 'org.springframework.boot' version '3.3.1'
     id 'io.spring.dependency-management' version '1.1.5'
 }
@@ -41,7 +44,7 @@ tasks.withType(JavaCompile).configureEach {
 }
 
 sourceSets {
-    main.java.srcDirs += [ generated ]
+    main.java.srcDirs += [generated]
 }
 
 clean {
