+++ domain/src/testFixtures/java/org/meotppo/webti/fixture/StatisticFixture.java
